backend/requirements.txt
fastapi
uvicorn
google-generativeai
python-multipart
pydantic
python-dotenv
sse-starlette
